requirements.txt
asgiref             3.4.1    
certifi             2021.5.30
cffi                1.14.6   
charset-normalizer  2.0.4    
cryptography        3.4.7    
defusedxml          0.7.1    
dj-database-url     0.5.0    
Django              3.2.6    
django-allauth      0.45.0   
django-environ      0.4.5    
django-heroku       0.3.1    
django-rest-auth    0.9.5    
djangorestframework 3.12.4   
environ             1.0      
gunicorn            20.1.0   
idna                3.2      
oauthlib            3.1.1    
pip                 20.0.2   
pkg-resources       0.0.0    
psycopg2            2.9.1    
psycopg2-binary     2.8.5    
pycparser           2.20     
PyJWT               2.1.0    
python3-openid      3.2.0    
pytz                2021.1   
requests            2.26.0   
requests-oauthlib   1.3.0    
setuptools          44.0.0   
six                 1.16.0   
sqlparse            0.4.1    
urllib3             1.26.6   
whitenoise          5.3.0    
